requirements.txt
# SDK RAG Agent — Python dependencies
anthropic>=0.25.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pymupdf>=1.24.0
sentence-transformers>=2.7.0
numpy>=1.26.0
pydantic>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
psycopg2-binary>=2.9.0
